+++ src/helpers/DynamicMeta.php
@@ -12,7 +12,9 @@
 namespace nystudio107\seomatic\helpers;
 
 use nystudio107\seomatic\models\Entity;
+use nystudio107\seomatic\models\jsonld\ContactPoint;
 use nystudio107\seomatic\models\jsonld\LocalBusiness;
+use nystudio107\seomatic\models\jsonld\Organization;
 use nystudio107\seomatic\Seomatic;
 use nystudio107\seomatic\models\jsonld\BreadcrumbList;
 use nystudio107\seomatic\models\jsonld\Thing;
@@ -23,6 +25,7 @@ use craft\base\Element;
 use craft\errors\SiteNotFoundException;
 use craft\helpers\DateTimeHelper;
 use craft\helpers\Json;
+use craft\web\twig\variables\Paginate;
 
 use yii\base\Exception;
 use yii\base\InvalidConfigException;
@@ -39,6 +42,39 @@ class DynamicMeta
 
     // Static Methods
     // =========================================================================
+
+    /**
+     * Paginate based on the passed in Paginate variable as returned from the
+     * Twig {% paginate %} tag: https://docs.craftcms.com/v3/templating/tags/paginate.html#the-pageInfo-variable
+     *
+     * @param Paginate $pageInfo
+     */
+    public static function paginate(Paginate $pageInfo)
+    {
+        if ($pageInfo !== null) {
+            // Set the current page
+            $url = $pageInfo->getPageUrl($pageInfo->currentPage);
+            if (!empty($url)) {
+                Seomatic::$seomaticVariable->meta->canonicalUrl = $url;
+            }
+            // Set the previous URL
+            $url = $pageInfo->getPrevUrl();
+            if (!empty($url)) {
+                $metaTag = Seomatic::$plugin->link->create([
+                    'rel' => 'prev',
+                    'href' => $url,
+                ]);
+            }
+            // Set the next URL
+            $url = $pageInfo->getNextUrl();
+            if (!empty($url)) {
+                $metaTag = Seomatic::$plugin->link->create([
+                    'rel' => 'next',
+                    'href' => $url,
+                ]);
+            }
+        }
+    }
 
     /**
      * Include any headers for this request
@@ -105,10 +141,12 @@ class DynamicMeta
             $jsonLd = Seomatic::$plugin->jsonLd->get('identity');
             if ($jsonLd !== null) {
                 self::addOpeningHours($jsonLd, $metaSiteVars->identity);
+                self::addContactPoints($jsonLd, $metaSiteVars->identity);
             }
             $jsonLd = Seomatic::$plugin->jsonLd->get('creator');
             if ($jsonLd !== null) {
                 self::addOpeningHours($jsonLd, $metaSiteVars->creator);
+                self::addContactPoints($jsonLd, $metaSiteVars->creator);
             }
         }
     }
@@ -145,17 +183,43 @@ class DynamicMeta
                     }
                 }
                 if ($openTime && $closeTime) {
-                    $spec = [
-                        'type'      => 'OpeningHoursSpecification',
-                        'opens'     => $openTime,
-                        'closes'    => $closeTime,
+                    $hours = Seomatic::$plugin->jsonLd->create([
+                        'type' => 'OpeningHoursSpecification',
+                        'opens' => $openTime,
+                        'closes' => $closeTime,
                         'dayOfWeek' => [$days[$index]],
-                    ];
-                    $openingHours[] = $spec;
+                    ], false);
+                    $openingHours[] = $hours;
                 }
                 $index++;
             }
             $jsonLd->openingHoursSpecification = $openingHours;
+        }
+    }
+
+    /**
+     * Add the ContactPoint to the $jsonLd based on the Entity settings
+     *
+     * @param MetaJsonLd $jsonLd
+     * @param Entity     $entity
+     */
+    public static function addContactPoints(MetaJsonLd $jsonLd, Entity $entity)
+    {
+        if ($jsonLd instanceof Organization && $entity !== null) {
+            /** @var Organization $jsonLd */
+            $contactPoints = [];
+            if ($entity->organizationContactPoints !== null && \is_array($entity->organizationContactPoints)) {
+                foreach ($entity->organizationContactPoints as $contacts) {
+                    /** @var ContactPoint $contact */
+                    $contact = Seomatic::$plugin->jsonLd->create([
+                        'type' => 'ContactPoint',
+                        'telephone' => $contacts['telephone'],
+                        'contactType' => $contacts['contactType'],
+                    ], false);
+                    $contactPoints[] = $contact;
+                }
+            }
+            $jsonLd->contactPoint = $contactPoints;
         }
     }
 
@@ -184,8 +248,8 @@ class DynamicMeta
         }
         /** @var  $crumbs BreadcrumbList */
         $crumbs = Seomatic::$plugin->jsonLd->create([
-            'type'        => 'BreadcrumbList',
-            'name'        => 'Breadcrumbs',
+            'type' => 'BreadcrumbList',
+            'name' => 'Breadcrumbs',
             'description' => 'Breadcrumbs list',
         ]);
         /** @var Element $element */
@@ -200,8 +264,8 @@ class DynamicMeta
             }
             $listItem = MetaJsonLd::create('ListItem', [
                 'position' => $position,
-                'item'     => [
-                    '@id'  => $id,
+                'item' => [
+                    '@id' => $id,
                     'name' => $element->title,
                 ],
             ]);
@@ -209,8 +273,8 @@ class DynamicMeta
         } else {
             $crumbs->itemListElement[] = MetaJsonLd::create('ListItem', [
                 'position' => $position,
-                'item'     => [
-                    '@id'  => $siteUrl,
+                'item' => [
+                    '@id' => $siteUrl,
                     'name' => 'Homepage',
                 ],
             ]);
@@ -243,8 +307,8 @@ class DynamicMeta
                 }
                 $crumbs->itemListElement[] = MetaJsonLd::create('ListItem', [
                     'position' => $position,
-                    'item'     => [
-                        '@id'  => $id,
+                    'item' => [
+                        '@id' => $id,
                         'name' => $element->title,
                     ],
                 ]);
@@ -265,9 +329,9 @@ class DynamicMeta
         // Add the x-default hreflang
         $siteLocalizedUrl = $siteLocalizedUrls[0];
         $metaTag = Seomatic::$plugin->link->create([
-            'rel'      => 'alternate',
+            'rel' => 'alternate',
             'hreflang' => ['x-default'],
-            'href'     => [$siteLocalizedUrl['url']],
+            'href' => [$siteLocalizedUrl['url']],
         ]);
         // Add the alternate language link rel's
         if (\count($siteLocalizedUrls) > 1) {
@@ -324,6 +388,26 @@ class DynamicMeta
     public static function getLocalizedUrls(string $uri = null, int $siteId = null): array
     {
         $localizedUrls = [];
+        // Set the pagination URL params, if they exist
+        $urlParams = null;
+        $pageTrigger = Craft::$app->getConfig()->getGeneral()->pageTrigger;
+        if (!\is_string($pageTrigger) || $pageTrigger === '') {
+            $pageTrigger = 'p';
+        }
+        // Is this query string-based pagination?
+        if ($pageTrigger[0] === '?') {
+            $pageTrigger = trim($pageTrigger, '?=');
+        }
+        // Avoid conflict with the path param
+        $pathParam = Craft::$app->getConfig()->getGeneral()->pathParam;
+        if ($pageTrigger === $pathParam) {
+            $pageTrigger = $pathParam === 'p' ? 'pg' : 'p';
+        }
+        $pageTriggerValue = Craft::$app->getRequest()->getParam($pageTrigger);
+        if ($pageTriggerValue !== null) {
+            $urlParams = [];
+            $urlParams[$pageTrigger] = $pageTriggerValue;
+        }
         // Get the request URI
         if ($uri === null) {
             try {
@@ -369,7 +453,7 @@ class DynamicMeta
                 $url = ($url === '__home__') ? '' : $url;
             } else {
                 try {
-                    $url = $site->hasUrls ? UrlHelper::siteUrl($requestUri, null, null, $site->id)
+                    $url = $site->hasUrls ? UrlHelper::siteUrl($requestUri, $urlParams, null, $site->id)
                         : Craft::$app->getSites()->getPrimarySite()->baseUrl;
                 } catch (SiteNotFoundException $e) {
                     $url = '';
@@ -382,7 +466,7 @@ class DynamicMeta
             $url = $url ?? '';
             if (!UrlHelper::isAbsoluteUrl($url)) {
                 try {
-                    $url = UrlHelper::siteUrl($url, null, null, $site->id);
+                    $url = UrlHelper::siteUrl($url, $urlParams, null, $site->id);
                 } catch (Exception $e) {
                     $url = '';
                     Craft::error($e->getMessage(), __METHOD__);
@@ -394,10 +478,10 @@ class DynamicMeta
             $hreflangLanguage = strtolower($hreflangLanguage);
             $hreflangLanguage = str_replace('_', '-', $hreflangLanguage);
             $localizedUrls[] = [
-                'id'               => $site->id,
-                'language'         => $language,
+                'id' => $site->id,
+                'language' => $language,
                 'hreflangLanguage' => $hreflangLanguage,
-                'url'              => $url,
+                'url' => $url,
             ];
         }